requirements.txt
anyio==4.6.2.post1; python_version >= '3.9'
asgiref==3.8.1; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dj-database-url==2.3.0
django==4.2.16; python_version >= '3.8'
django-choices-field==2.3.0; python_version >= '3.8' and python_version < '4.0'
django-cors-headers==4.4.0; python_version >= '3.8'
django-multiselectfield==0.1.13
exceptiongroup==1.2.2; python_version < '3.11'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.4
idna==3.10; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
psycopg[binary]==3.2.3; python_version >= '3.8'
psycopg-binary==3.2.3
pyjwt==2.9.0; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.2
requests==2.32.3; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.11'
urllib3==2.2.3; python_version >= '3.8'
uvicorn[standard]==0.30.6; python_version >= '3.8'
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1